requirements.txt
alembic==1.6.5
aniso8601==9.0.1
astroid==2.6.6
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.8
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pylint==2.9.6
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-editor==1.0.4
python-jose==3.3.0
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1